chore(clippy): apply auto-fix suggestions across workspace [I.D D2]
cargo clippy --workspace --all-targets --fix + bin "tasty" 보조 패스 적용:
- collapsible_if (180건) → Rust 2024 let-chain 통합
- needless_borrows_for_generic_args (30건)
- redundant_field_names (12건)
- unnecessary_cast / map_or / redundant_closure / manual_range_contains 등

수동 보정 2건:
- src/file/format/registry/user_edit.rs: filter_map_bool_then 가 매크로
  metavariable 잔여로 invalid 코드 생성 → 직접 filter+map 으로 풀어 씀
- crates/tasty-themes/src/plugin_defaults.rs: auto-fix 가 SurfaceTheme 임포트
  를 unused 로 잘못 제거 → 테스트 사용처 살아있어 복구

cargo build / test --no-run 통과 확인.

+++ src/adapters/ui/input/click_cursor.rs
@@ -54,7 +54,7 @@ impl EditableRegion {
 
         let (cols, rows) = terminal.surface().dimensions();
         let (cursor_col, cursor_row) = terminal.surface().cursor_position();
-        let cursor_row = cursor_row as usize;
+        let cursor_row = cursor_row;
         let screen_lines = terminal.surface().screen_lines();
 
         // Walk upward from cursor_row to find the first row of the editable region.
